backend/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
djangorestframework-simplejwt
Pillow
python-decouple
mysqlclient
drf-spectacular
django-extensions
Faker
factory-boy
redis
django-redis
celery==5.4.0
django-celery-email==3.0.0
gunicorn
django-filter>=23.1
whitenoise
django-storages